Added snapshot.NewIndexedBatch
Adds a new method to the Snapshot type to create a new IndexedBatch using the snapshot. This means that all reads on the IndexedBatch will either come from:
A) The regular batch index
B) The snapshot state

This allows us to build snapshot isolation mechanics into the Batch store, which already gives us Atomicity and Durability.
Only one step away from full ACID transactions for pebble (note: Consistency is the last piece of the puzzle).